Follow screenOrientation
Apps that don't allow me to re-orient them are one of my pet-peeves and forcing
apps to follow screenOrientation anyways through an overlay works in 95/100
cases.

Here, it wasn't as trivial though because changing the orientation rotates the
app and therefore the compass. Landscape, inverse portrait and inverse landscape
need it to be rotated by -90, -180 and -270 degrees respectively to compensate
for this effect.

The implementation is not pretty but it's the best API for that kind of thing I
could find.